+++ train_hybrid_transformer.py
@@ -6,7 +6,8 @@ from sklearn.datasets import fetch_20newsgroups
 from sklearn.feature_extraction.text import CountVectorizer
 import matplotlib.pyplot as plt
 
-from src.hybrid_transformer import ClassicalTransformer, HybridTransformer
+from src.classical_transformer import ClassicalTransformer
+from src.hybrid_transformer import HybridTransformer
 
 
 class TextDataset(Dataset):
